requirement.txt
streamlit
python-dotenv
sentence-transformers
chromadb
pdfplumber
beautifulsoup4
requests
openai
pillow
pytesseract
